Move the class OpaqueTQ3Object here from E3ClassTree.h
Added a macro to easily set up the enum at the start of each class to record the class's depth in the Quesa object hierarchy, the class's type and also the parent's type (the latter is not yet used)

Called the macro in all the classes in E3Main except the root class which has to do things by hand.

Added the enum for the root class (OpaqueTQ3Object)

Made OpaqueTQ3Object::IsObjectVaild into an inline which gives a speed improvement.


git-svn-id: svn://svn.code.sf.net/p/quesa/code/trunk/quesa@1900 a5a2676d-45bc-4c7e-904d-7c40250887ca